# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SIC API
  version: 0.1.0
  description: Implementation of accounting system based in hordak and following SIC
    (Andrisani) definitions.
  contact:
    name: gastong256
paths:
...
  /api/v1/admin/registration-code/:
    get:
      operationId: admin_get_registration_code
      summary: Get current registration code
      tags:
      - admin
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationCodeInfo'
          description: ''
  /api/v1/admin/registration-code/rotate/:
    post:
      operationId: rotate_registration_code
      summary: Rotate registration code
      tags:
      - admin
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationCodeInfo'
          description: ''
  /api/v1/admin/users/:
    get:
      operationId: admin_list_users
      summary: List users
      parameters:
      - in: query
        name: all
        schema:
          type: boolean
        description: Return the full list in one response.
      - in: query
        name: page
        schema:
          type: integer
        description: Pagination page number.
      - in: query
        name: role
        schema:
          type: string
          enum:
          - admin
          - student
          - teacher
        description: Filter users by role.
      - in: query
        name: search
        schema:
          type: string
        description: Case-insensitive username search.
      - in: query
        name: summary
        schema:
          type: string
          enum:
          - selector
        description: Return a lightweight item shape for selectors.
      tags:
      - admin
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListPaginated'
          description: ''
  /api/v1/admin/users/{user_id}/role/:
    patch:
      operationId: admin_update_user_role
      summary: Update user role
      parameters:
      - in: path
        name: user_id
        schema:
          type: integer
        required: true
      tags:
      - admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedUserRoleUpdate'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
  /api/v1/auth/me/:
    get:
      operationId: get_me
      summary: Get current user
      parameters:
      - in: query
        name: include
        schema:
          type: string
        description: 'Comma-separated extra sections to include. Supported values:
          companies, capabilities, registration_code.'
      tags:
      - auth
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
          description: ''
    patch:
      operationId: update_me
      summary: Update current user profile
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedUserUpdate'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
  /api/v1/auth/register/:
    post:
      operationId: register_student
      summary: Register new student user
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: ''
        '400':
          description: Validation error
        '429':
          description: Rate limited or cooldown active
...
  /api/v1/companies/:
    get:
      operationId: list_companies
      summary: List companies
      parameters:
      - in: query
        name: all
        schema:
          type: boolean
        description: Return the full list in one response.
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - in: query
        name: page
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: summary
        schema:
          type: string
          enum:
          - selector
        description: Return a lightweight item shape for selectors.
      tags:
      - companies
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCompanyList'
          description: ''
    post:
      operationId: create_company
      summary: Create a company
      tags:
      - companies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCreate'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
        '400':
          description: Validation error
...
  /api/v1/teacher/registration-code/:
    get:
      operationId: teacher_get_registration_code
      summary: Get current registration code
      tags:
      - teacher
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationCodeInfo'
          description: ''
...
components:
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 255
        description:
          type: string
          description: Optional business description for a more realistic simulation.
        tax_id:
          type: string
          description: Simulated CUIT (optional).
          maxLength: 20
        owner_username:
          type: string
          readOnly: true
        account_count:
          type: integer
          readOnly: true
        books_closed_until:
          type: string
          format: date
          readOnly: true
          nullable: true
          description: Entries on or before this date are locked.
        is_demo:
          type: boolean
          readOnly: true
          description: Marks a seeded demo company shared across the system.
        is_read_only:
          type: boolean
          readOnly: true
          description: If true, accounting writes and metadata updates are blocked.
        has_opening_entry:
          type: boolean
          readOnly: true
        accounting_ready:
          type: boolean
          readOnly: true
        opening_entry_id:
          type: integer
          nullable: true
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - account_count
      - accounting_ready
      - books_closed_until
      - created_at
      - has_opening_entry
      - id
      - is_demo
      - is_read_only
      - name
      - opening_entry_id
      - owner_username
      - updated_at
    CompanyCreate:
      type: object
      properties:
        name:
          type: string
          description: Company name.
          maxLength: 255
        description:
          type: string
          default: ''
          description: Optional business description.
        tax_id:
          type: string
          default: ''
          description: Simulated CUIT (optional).
          maxLength: 20
        opening_entry:
          allOf:
          - $ref: '#/components/schemas/CompanyOpeningEntry'
          description: Optional balanced opening entry to create the company with
            initial capital/assets.
      required:
      - name
    CompanyOpeningAccountItem:
      type: object
      properties:
        name:
          type: string
          description: Movement-account name to create or reuse for this company.
          maxLength: 255
        parent_code:
          type: string
          description: Parent colectiva code in format X.XX (e.g. 1.01).
          maxLength: 4
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          description: Positive amount.
      required:
      - amount
      - name
      - parent_code
    CompanyOpeningEntry:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Accounting date of the opening entry.
        inventory_kind:
          allOf:
          - $ref: '#/components/schemas/InventoryKindEnum'
          default: INITIAL
          description: |-
            Whether the opening is based on the initial inventory or a later general inventory.

            * `INITIAL` - Inventario Inicial
            * `GENERAL` - Inventario General
        source_ref:
          type: string
          default: ''
          description: Optional reference, e.g. "APERTURA-001" or "INV-GRAL-001".
          maxLength: 100
        assets:
          type: array
          items:
            $ref: '#/components/schemas/CompanyOpeningAccountItem'
          description: Asset items that will be posted to Debe.
        liabilities:
          type: array
          items:
            $ref: '#/components/schemas/CompanyOpeningAccountItem'
          description: Liability items that will be posted to Haber.
      required:
      - assets
      - date
    CompanySelector:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          readOnly: true
        owner_username:
          type: string
          readOnly: true
        is_demo:
          type: boolean
          readOnly: true
          description: Marks a seeded demo company shared across the system.
        is_read_only:
          type: boolean
          readOnly: true
          description: If true, accounting writes and metadata updates are blocked.
        has_opening_entry:
          type: boolean
          readOnly: true
        accounting_ready:
          type: boolean
          readOnly: true
      required:
      - accounting_ready
      - has_opening_entry
      - id
      - is_demo
      - is_read_only
      - name
      - owner_username
...
    InventoryKindEnum:
      enum:
      - INITIAL
      - GENERAL
      type: string
      description: |-
        * `INITIAL` - Inventario Inicial
        * `GENERAL` - Inventario General
...
    MeResponse:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          readOnly: true
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
        email:
          type: string
          format: email
          readOnly: true
          title: Email address
        first_name:
          type: string
          readOnly: true
        last_name:
          type: string
          readOnly: true
        is_staff:
          type: boolean
          readOnly: true
          title: Staff status
          description: Designates whether the user can log into this admin site.
        role:
          type: string
          readOnly: true
        course_id:
          type: integer
          nullable: true
          readOnly: true
        date_joined:
          type: string
          format: date-time
          readOnly: true
        companies:
          type: array
          items:
            $ref: '#/components/schemas/CompanySelector'
        capabilities:
          $ref: '#/components/schemas/UserCapabilities'
        registration_code:
          allOf:
          - $ref: '#/components/schemas/RegistrationCodeInfo'
          nullable: true
      required:
      - course_id
      - date_joined
      - email
      - first_name
      - id
      - is_staff
      - last_name
      - role
      - username
    PaginatedCompanyList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Company'
...
    PatchedUserRoleUpdate:
      type: object
      properties:
        role:
          $ref: '#/components/schemas/RoleEnum'
    PatchedUserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
...
    RegistrationCodeInfo:
      type: object
      properties:
        code:
          type: string
        window_minutes:
          type: integer
        allow_previous_window:
          type: boolean
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
      required:
      - allow_previous_window
      - code
      - valid_from
      - valid_until
      - window_minutes
    RoleEnum:
      enum:
      - admin
      - teacher
      - student
      type: string
      description: |-
        * `admin` - Admin
        * `teacher` - Teacher
        * `student` - Student
...
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          readOnly: true
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
        email:
          type: string
          format: email
          readOnly: true
          title: Email address
        first_name:
          type: string
          readOnly: true
        last_name:
          type: string
          readOnly: true
        is_staff:
          type: boolean
          readOnly: true
          title: Staff status
          description: Designates whether the user can log into this admin site.
        role:
          type: string
          readOnly: true
        course_id:
          type: integer
          nullable: true
          readOnly: true
        date_joined:
          type: string
          format: date-time
          readOnly: true
      required:
      - course_id
      - date_joined
      - email
      - first_name
      - id
      - is_staff
      - last_name
      - role
      - username
    UserCapabilities:
      type: object
      properties:
        can_manage_courses:
          type: boolean
        can_manage_visibility:
          type: boolean
        can_view_registration_code:
          type: boolean
        can_manage_roles:
          type: boolean
      required:
      - can_manage_courses
      - can_manage_roles
      - can_manage_visibility
      - can_view_registration_code
    UserListPaginated:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          nullable: true
        previous:
          type: string
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/User'
      required:
      - count
      - next
      - previous
      - results
    UserRegister:
      type: object
      properties:
        username:
          type: string
          maxLength: 150
        password:
          type: string
          writeOnly: true
          minLength: 8
        password_confirm:
          type: string
          writeOnly: true
          minLength: 8
        email:
          type: string
          format: email
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        registration_code:
          type: string
          maxLength: 32
      required:
      - password
      - password_confirm
      - registration_code
      - username
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT